[NV] Use FP8 conversion intrinsics, when available
PTX "cvt" instruction supports converting to/from FP8 types.
The NV hardware supports E4M3FN and E5M2 types.
This PR updates the MLIR emitter to use this instruction instead of emitting
a long sequence of operations (this matters in compute-bound FP8 kernels).

The NVVM intrinsic allows converting two FP8 values with a single instruction, but
as the emitter is elementwise, only one of the inputs is used. This is wasteful,
but still much faster than emitting the sequence of instructions.

Before ptx 7.8 (cuda 11.8), the instruction is not supported.
Starting with ptx 8.1 (cuda 12.1), the instruction is supported for sm89+.
Between those versions, the instruction is supported for sm90+, thas is, if trying to compile
on Ada (sm89) with cuda version < 12.1, the ptxas will complain..

Reference:
https://docs.nvidia.com/cuda/parallel-thread-execution/index.html#data-movement-and-conversion-instructions-cvt
(see "PTX ISA Notes" and "Target ISA Notes").